edit : 타입 수정

+++ Back-end/api-module/src/main/java/com/ico/api/service/License/LicenseServiceImpl.java
@@ -199,7 +199,7 @@ public class LicenseServiceImpl implements LicenseService{
                 continue;
             }
 
-            license.setRating((byte) rating.intValue());
+            license.setRating(rating.byteValue());
             studentLicenseRepository.save(license);
         }
     }
